ㅁㅁㅁ

+++ study-project/src/main/java/com/study_project/web/soju/controller/SojuController.java
@@ -1,6 +1,5 @@
 package com.study_project.web.soju.controller;
 
-import java.io.File;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -22,7 +21,10 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 
 import com.study_project.web.soju.model.Soju;
+import com.study_project.web.soju.model.SojuComment;
+import com.study_project.web.soju.service.SojuCommentService;
 import com.study_project.web.soju.service.SojuService;
+import com.study_project.web.user.model.User;
 import com.study_project.web.util.FileUtil;
 import com.study_project.web.util.Pager;
 
@@ -35,6 +37,9 @@ public class SojuController {
 	
 	@Autowired
 	private SojuService sojuService;
+	
+	@Autowired
+	private SojuCommentService sojuCommentService;
 	
 	//글목록
 	@RequestMapping(value="soju.html", method = RequestMethod.GET)
@@ -81,10 +86,26 @@ public class SojuController {
 	
 	//상세보기 입력폼전환
 	@RequestMapping(value="/{idx}", method = RequestMethod.GET)
-	public String detailForm (@PathVariable("idx") Integer idx, Model model) throws Exception {
+	public String detailForm (@PathVariable("idx") Integer idx, Model model, HttpSession session) throws Exception {
 		logger.info("[ welcome sojuDetailForm ]");
+		User user = (User)session.getAttribute("user");
+		if (user == null) {
+			String loginAdvice = "로그인 후 이용하실 수 있습니다. 로그인 해주세요.";
+			model.addAttribute("loginAdvice", loginAdvice);
+			return "/user/login";
+		}
+		String loginUserCheck = user.getName();
+		
 		Soju soju = sojuService.getSoju(idx);
+		List<Soju> sojuFileList = sojuService.selectSojuFile(idx);
+		List<SojuComment> sojuCommentList = sojuCommentService.selectSojuCommentList(idx);
+		Integer sojuCommentCount = sojuCommentService.count(idx);
+		
+		model.addAttribute("sojuCommentCount", sojuCommentCount);
+		model.addAttribute("sojuCommentList", sojuCommentList);
 		model.addAttribute("soju", soju);
+		model.addAttribute("sojuFile",sojuFileList);
+		model.addAttribute("loginUserCheck", loginUserCheck);
 		return "/soju/sojuDetail";
 	}
 	
@@ -128,34 +149,33 @@ public class SojuController {
 				response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Expected multipart requeest");
 				return null;
 			}
+			String thumbnailO_name = multipart.getOriginalFilename();
+			String thumbnailS_name = fileUtil.makeServerFileName(fileUtil.getStrNowTime(), thumbnailO_name);
+			soju.setThumbnail(thumbnailS_name);
 			Integer insertSojuResult = sojuService.insertSoju(soju);
 			if (insertSojuResult == 1) {
-				String thumbnailO_name = multipart.getOriginalFilename();
-				String thumbnailS_name = fileUtil.makeServerFileName(fileUtil.getStrNowTime(), thumbnailO_name);
 				fileUtil.uploadFile(multipart, fileUtil.SojuThumbnailPath, thumbnailS_name);
-				soju.setThumbnail(thumbnailS_name);
-				
 			}
-			
 			MultipartHttpServletRequest multipartRequest = (MultipartHttpServletRequest) request;
 			for (String fileName : multipartRequest.getFileMap().keySet()) {
 				for (MultipartFile file : multipartRequest.getFiles(fileName)) {
 					String O_name = file.getOriginalFilename();
-					//시간+파일이름 이름을 만듬
-					String transFileName = fileUtil.makeServerFileName(fileUtil.getStrNowTime(), O_name);
-					soju.setOriginal_name(O_name);
-					soju.setTrans_name(transFileName);
-					soju.setFile_path(fileUtil.SojuproductImagePath);
-					soju.setFile_size(file.getSize());
-					soju.setSoju_idx(soju.getIdx());
-					soju.setContent_type(fileUtil.fileTypeTrans(O_name));
-					/*File destination = File.createTempFile("file", transFileName, new File(fileUtil.SojuproductImagePath));*/
-					fileUtil.uploadFile(file, fileUtil.SojuproductImagePath, transFileName);
-					sojuService.insertSojuFile(soju);
+					if (file.getSize() != 0) {
+						if (O_name == thumbnailO_name) continue;
+						//시간+파일이름 이름을 만듬
+						String transFileName = fileUtil.makeServerFileName(fileUtil.getStrNowTime(), O_name);
+						soju.setOriginal_name(O_name);
+						soju.setTrans_name(transFileName);
+						soju.setFile_path(fileUtil.SojuproductImagePath);
+						soju.setFile_size(file.getSize());
+						soju.setSoju_idx(soju.getIdx());
+						soju.setContent_type(fileUtil.fileTypeTrans(O_name));
+						fileUtil.uploadFile(file, fileUtil.SojuproductImagePath, transFileName);
+						sojuService.insertSojuFile(soju);
+					}
+					
 				}
 			}
-			
-			
 			logger.info("[ sojuReg toString ] " + soju.toString());
 		} catch(Exception e){
 			e.printStackTrace();
@@ -171,12 +191,83 @@ public class SojuController {
 	
 	//삭제 로직 수행
 	@RequestMapping(value="/board/{idx}", method = RequestMethod.DELETE)
-	public String delete (@PathVariable("idx") Integer idx) throws Exception {
+	public String delete (@PathVariable("idx") Integer idx, Soju soju1) throws Exception {
 		logger.info("[ welcome sojuDelete ]");
-		sojuService.deleteSoju(idx);
+		FileUtil fileUtil = new FileUtil();
+		
+		List<Soju> sojuFileList = sojuService.selectSojuFile(idx);
+		
+		if (idx != null) {
+			boolean fileDelete = false;
+			//썸네일 이미지 삭제
+			Soju sojuDetail = sojuService.getSoju(idx);
+			soju1.setThumbnail(sojuDetail.getThumbnail());
+			System.out.println("썸네일이 들어왔나 = " + soju1.getThumbnail());
+			boolean thumbnailDelete = fileUtil.deleteFile(fileUtil.SojuThumbnailPath, soju1.getThumbnail());
+			if (thumbnailDelete == true) {
+				//상품소개(다중) 삭제
+				for (Soju soju : sojuFileList) {
+					System.out.println("fileidx == " + soju.getFile_idx());
+					logger.info("logger file idx tostring = " + sojuFileList.get(0).getFile_idx());
+					fileDelete =fileUtil.deleteFile(fileUtil.SojuproductImagePath, soju.getTrans_name());
+				}
+			}
+			if (fileDelete == true) {
+				sojuService.deleteSoju(idx);
+			}
+		}
 		return "redirect:/soju/soju.html";
 	}
 	
+	//상품평 입력
+	@RequestMapping(value="/board/{idx}/comment", method = RequestMethod.POST)
+	public String commentInsert (@PathVariable("idx") Integer idx,Model model, SojuComment sojuComment,HttpSession session) throws Exception {
+		logger.info(" [ welcome soju Insert Comment ] ");
+		try {
+			User user = (User)session.getAttribute("user");
+			sojuComment.setId(user.getName());
+			if (user.getName() == null) {
+				return "/user/login.html";
+			}
+			sojuComment.setSoju_idx(idx);
+			sojuCommentService.insertSojuComment(sojuComment);
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+//		return detailForm(idx,model,session);
+		return "redirect:/soju/{idx}";
+	}
+	
+	//상품평 삭제
+	@RequestMapping(value="/board/{idx}/comment/{commentidx}", method = RequestMethod.POST)
+	public String commentDelete (@PathVariable("idx") Integer idx
+			, @PathVariable("commentidx") Integer commentidx, Model model, HttpSession session) throws Exception {
+		logger.info(" [ welcome soju Delete Comment ] ");
+		SojuComment sojuComment = new SojuComment();
+		sojuComment.setIdx(commentidx);
+		sojuComment.setSoju_idx(idx);
+		sojuCommentService.deleteSojuComment(sojuComment);
+		
+//		return detailForm(idx,model,session);
+		return "redirect:/soju/{idx}";
+	}
+	
+	@RequestMapping(value="/board/{idx}/comment/{commentidx}/modify/", method = RequestMethod.POST)
+	public String commentModify (@PathVariable("idx") Integer idx
+			, @PathVariable("commentidx") Integer commentidx
+			, @RequestParam("edit") String edit
+			, Model model, HttpSession session, SojuComment sojuComment) throws Exception {
+		logger.info(" [ welcome soju Modify Comment ] ");
+		logger.info(" [ SojuComment Model > ]" + sojuComment);
+		System.out.println("에딧 = " + edit);
+		sojuComment.setContents(edit);
+		sojuComment.setIdx(commentidx);
+		sojuComment.setSoju_idx(idx);
+		sojuCommentService.updateSojuComment(sojuComment);
+		
+//		return detailForm(idx,model,session);
+		return "redirect:/soju/{idx}";
+	}
